+++ contracts/test-libraries/utils/TestLib_EthUtils.sol
@@ -85,7 +85,7 @@ contract TestLib_EthUtils {
 
     function getAddressForCREATE(
         address _creator,
-        uint256 _nonce
+        uint64 _nonce
     )
         public
         pure
